requirements.txt
fastapi
pydantic
uvicorn[standard]
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
openai
bs4
faiss-cpu
python-dotenv
colorama
tqdm
tiktoken
transformers
torch
sentence-transformers
typing-extensions
numpy
accelerate
chromadb
websockets
python-multipart
boto3
PyPDF2
pycryptodome
unidecode
pytesseract
pdf2image
pillow